+++ lib/src/apollovm_base.dart
@@ -44,7 +44,7 @@ import 'languages/wasm/wasm_runner.dart';
 /// The Apollo VM.
 class ApolloVM implements VMTypeResolver {
   // ignore: non_constant_identifier_names
-  static final String VERSION = '0.1.29';
+  static final String VERSION = '0.1.30';
 
   static int _idCount = 0;